Move v1 out to a different repo. Lints

+++ src/memeasm/assembler.cpp
@@ -2,6 +2,7 @@
 #include "memeasm/assembler.h"
 
 #include <cinttypes>
+#include <utility>
 
 #include "boost/property_tree/json_parser.hpp"
 #include "boost/property_tree/ptree.hpp"
@@ -25,7 +26,8 @@ std::vector<std::string> getLines(const std::string& data) {
   std::vector<std::string> lines;
   while (std::getline(ss, line)) {
     int last_idx = 0;
-    while (last_idx < line.size() && line[last_idx] != ';') last_idx++;  // Remove comments.
+    while (last_idx < static_cast<int>(line.size()) && line[last_idx] != ';')
+      last_idx++;  // Remove comments.
     const auto& s = trim(line.substr(0, last_idx), " \t\n");
     lines.push_back(s);  // Include blank lines so line numbers are accurate.
   }
@@ -46,7 +48,7 @@ Assembler::Assembler(const std::string& model_json) {
     mnemonic.opcode = kv.second.get<int>("opcode");
     mnemonic.imm_relative = kv.second.get<bool>("imm_relative");
     std::string mnemonic_rx = PREAMBLE_RX;
-    for (int i = 0; i < mnemonic_str.size(); ++i) {
+    for (int i = 0; i < static_cast<int>(mnemonic_str.size()); ++i) {
       if (mnemonic_str[i] == '%') {
         verify(i < static_cast<int>(mnemonic_str.size()) - 2, "ill-formed ksm model (BUG)");
         i += 2;  // Skip % and number.
@@ -95,7 +97,7 @@ std::vector<uint32_t> Assembler::assemble(const std::string& data) {
 void Assembler::assembleInternal(bool first_pass) {
   std::regex label_rx(std::string(PREAMBLE_RX) + LABEL_RX + POSTAMBLE_RX);
 
-  for (int lnum = 0; lnum < lines_.size(); ++lnum) {
+  for (int lnum = 0; lnum < static_cast<int>(lines_.size()); ++lnum) {
     const std::string& line = lines_[lnum];
     if (line.empty()) continue;
 
@@ -127,8 +129,8 @@ uint32_t Assembler::convertMnemonicStringToOpword(
 
     auto opword = uint32_t(mnemonic.opcode);
     int reg_count = 0;
-    for (int i = 0; i < mnemonic.params.size(); ++i) {
-      verify(i + 1 < sm.size(), "%d: missing parameter %s", lnum, line.c_str());
+    for (int i = 0; i < static_cast<int>(mnemonic.params.size()); ++i) {
+      verify(i + 1 < static_cast<int>(sm.size()), "%d: missing parameter %s", lnum, line.c_str());
       const auto& pstr = sm[i + 1].str();
       switch (mnemonic.params[i]) {
       case Parameter::REGISTER: {
