+++ src/memelang/tokeniser.cpp
@@ -56,11 +56,11 @@ Tokeniser::Tokeniser(const FileContents* cts) : cts_(cts) { bug_unless(cts_); }
 std::vector<Tok> Tokeniser::tokenise() {
   toks_.clear();
   const auto& data = cts_->data();
-  for (int i = 0; i < data.size(); ++i)
+  for (int i = 0; i < static_cast<int>(data.size()); ++i)
     verify(isprint(data[i]) || data[i] == '\n', "unprintable character '%c' at %s", data[i],
         cts_->fpos(i).c_str());
 
-  while (idx_ < data.size()) {
+  while (idx_ < static_cast<int>(data.size())) {
     if (startsNewToken(data[idx_]) || atCompleteToken()) pushCurrentToken();
 
     if (!isspace(data[idx_]))
@@ -159,14 +159,14 @@ bool Tokeniser::atCompleteToken() {
 
 bool Tokeniser::isChar(char c, const char* msg) {
   const auto& data = cts_->data();
-  verify(idx_ < data.size(), msg);
+  verify(idx_ < static_cast<int>(data.size()), msg);
   return data[idx_] == c;
 }
 
 char Tokeniser::grabEscapedChar() {
   const auto& data = cts_->data();
   if (data[idx_] != '\\') return data[idx_++];
-  verify(++idx_ < data.size(), "unexpected EOF in escape sequence");
+  verify(++idx_ < static_cast<int>(data.size()), "unexpected EOF in escape sequence");
   idx_++;
   switch (data[idx_ - 1]) {
   case 'n': return '\n';
